strike_backend/requirements.txt
Django==4.2.25
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
gunicorn==23.0.0
